Simplify debug metadata: instead of array of same values, store single element
This change is valid, since in we want StackTraceElement to be as close
as possible to StackTraceElement of non-coroutine code.
Particularly, if the line is inside inlined function, it's className,
functionName and sourceFile are inline site ones, but lineNumber
is large, which is the way of representing inline function's line
numbers. The rest handled by the plugin, which allows the user to choose
between inline function body and inline site.